refactor: move data from lib to new data directory to improve folder structure
