+++ data/tech.tsx
